style: Comment out 'soup' cell to avoid endless scrolling
In this lesson, I learned how to use BeautifulSoup. In order
to find what I wanted in the 'soup', I displayed the whole
html in the notebook. Once I uploaded the file to GitHub,
I noticed the cell's huge output wouldn't collapse.
Therefore I opted to comment out the cell.